+++ include/mgard-x/DataRefactoring/MultiDimension/Correction/LinearProcessingKernel.hpp
@@ -539,9 +539,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     for (DIM d = 0; d < D; d++) {
       if (d != curr_dim_f && d != curr_dim_c && d != curr_dim_r) {
@@ -1001,9 +1001,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     for (DIM d = 0; d < D; d++) {
       if (d != curr_dim_f && d != curr_dim_c && d != curr_dim_r) {
@@ -1473,9 +1473,9 @@ public:
     tbz = R;
     tby = C;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     for (DIM d = 0; d < D; d++) {
       if (d != curr_dim_f && d != curr_dim_c && d != curr_dim_r) {
